sql: fix LOG of very large numbers
Previously a LOG of a large number would result in the server consuming
all CPU and hanging. This is because our LOG implementation would take 1/x
at a precision low enough that the division of the large x would result
in a 0. The algorithm would proceed to double it until it reached 0.5,
which obviously would never succeed. Fix this by using a high enough
precision to prevent 0 from being returned.

During the investigation of this bug, it was also discovered that if a
large number was near the vicinity of producing a 0, it would incorrectly
truncate some precision of the division, thus producing slightly incorrect
results to the LOG function.

This thus fixes various bugs and hangs in the POW, LOG, and LN builtins.

Fixes current RSG test failures.

Benchmark shows no change:

name          old time/op  new time/op  delta
DecimalLog-4   106µs ± 3%   105µs ± 1%  -1.26%  (p=0.029 n=10+10)
DecimalExp-4  18.6ms ±17%  18.4ms ± 8%    ~     (p=0.897 n=10+8)
DecimalPow-4  1.25ms ±18%  1.25ms ±21%    ~     (p=0.853 n=10+10)
