+++ pkg/util/decimal/decimal_test.go
@@ -317,6 +317,13 @@ func TestDecimalLog(t *testing.T) {
 		{"2", "0.6931471805599453"},
 		{"1234.56789", "7.1184763011977896"},
 		{"1234567898765432112.2763812", "41.6572527032084749"},
+		{"100000000000000000000000000000000", "73.6827229758094619"},
+		// Test the 1/x division to make sure it doesn't round to zero or truncate
+		// precision after the division.
+		{"123450000000000000000000000000000", "73.8933890056125590"},
+		{"1000000000000000000000000000000000", "75.9853080688035076"},
+		{"10000000000000000000000000000000000000000000000", "105.9189142777261015"},
+		{"1000002350000002340000000345354700000000764000009", "110.5240868137114339"},
 	}
 	testDecimalSingleArgFunc(t, Log, 16, tests)
 }
@@ -329,6 +336,13 @@ func TestDecimalLogDoubleScale(t *testing.T) {
 		{"2", "0.69314718055994530941723212145818"},
 		{"1234.56789", "7.11847630119778961310397607454138"},
 		{"1234567898765432112.2763812", "41.65725270320847492372271693721825"},
+		{"100000000000000000000000000000000", "73.68272297580946188857572654989965"},
+		// Test the 1/x division to make sure it doesn't round to zero or truncate
+		// precision after the division.
+		{"123450000000000000000000000000000", "73.89338900561255903040963826675629"},
+		{"1000000000000000000000000000000000", "75.98530806880350757259371800458402"},
+		{"10000000000000000000000000000000000000000000000", "105.91891427772610146482760691548075"},
+		{"1000002350000002340000000345354700000000764000009", "110.52408681371143392718404189196936"},
 	}
 	testDecimalSingleArgFunc(t, Log, 32, tests)
 }
